// apps/shared/crud-service/models/mapType.ts
export class ProductTypeMap
  implements CrudTypeMap
{
  aggregate: Prisma.ProductAggregateArgs;
  count: Prisma.ProductArgs;
  create: Prisma.ProductCreateArgs;
  delete: Prisma.ProductDeleteArgs;
  deleteMany: Prisma.ProductDeleteManyArgs;
  findFirst: Prisma.ProductFindFirstArgs;
  findMany: Prisma.ProductFindManyArgs;
  findUnique: Prisma.ProductFindUniqueArgs;
  update: Prisma.ProductUpdateArgs;
  updateMany: Prisma.ProductUpdateManyArgs;
  upsert: Prisma.ProductUpsertArgs;
}